pagination and others

+++ app/main/service/card_list_service.py
@@ -1,7 +1,11 @@
-from app.main import db
-from app.main.model.card_list import CardListModel
+from flask import request
 
-from app.main.service.user_service import get_a_user
+from app.main import db
+from app.main.config import PAGINATION_NUMBER
+from app.main.model.card_list import CardListModel
+from app.main.service.auth_service import Auth
+
+from app.main.service.user_service import get_a_user, user_has_card_list
 
 
 def save_new_card_list(data):
@@ -13,12 +17,12 @@ def save_new_card_list(data):
         }
         return response_object, 400
     else:
-        new_card = CardListModel(
+        new_card_list = CardListModel(
             title=data['title'],
         )
         try:
-            _save_changes(new_card)
-            return new_card.json(), 201
+            _save_changes(new_card_list)
+            return new_card_list.json(), 201
         except:
             error_object = {
                 'status': 'fail',
@@ -28,8 +32,8 @@ def save_new_card_list(data):
             return error_object, 500
 
 
-def get_all_lists():
-    return CardListModel.query.all()
+def get_all_lists_paginated(page=1):
+    return CardListModel.query.paginate(page, PAGINATION_NUMBER, error_out=False)
 
 
 def get_a_card_list_by_id(_id):
@@ -41,8 +45,14 @@ def get_a_card_list_by_title(title):
 
 
 def edit_a_card_list(_id, data):
-    print(data['title'])
     card_list = get_a_card_list_by_id(_id)
+    logged_user_id = Auth.get_logged_in_user_id(request)
+    user = user_has_card_list(_id, logged_user_id)
+    if not user:
+        return {'status': 'fail', 'message': 'UNAUTHORIZED'}, 401
+    print('user')
+    print(user)
+
     if not card_list:
         response_object = {
             'status': 'fail',
@@ -62,6 +72,13 @@ def edit_a_card_list(_id, data):
 
             _save_changes(card_list)
         return card_list.json(), 200
+
+
+def check_ownership(user_id):
+    logged_user_id = Auth.get_logged_in_user_id(request)
+    if user_id == logged_user_id:
+        return True
+    return False
 
 
 def assign_member(card_list_id, user_id, assign):
@@ -89,8 +106,13 @@ def assign_member(card_list_id, user_id, assign):
 
 
 def delete_card_list(card_list):
+    logged_user_id = Auth.get_logged_in_user_id(request)
+    user = user_has_card_list(card_list.id, logged_user_id)
+    if not user:
+        return {'status': 'fail', 'message': 'UNAUTHORIZED'}, 401
     db.session.delete(card_list)
     db.session.commit()
+    return {'status': 'success', 'message': 'Card list successfully deleted. '}, 200
 
 
 def _save_changes(data):
